Initial implementation based on the original implementation
The scheduled task has been originally implemented as a part of a local
plugin used on the site learn.moodle.net by Dan Poltawski back in 2014.

In order to allow the feature to be distributed and used on other sites,
I have refactored it into a stand-alone plugin.

The settings "Maximum auto-approved courses" and "Auto-reject extra
courses" are new. Originally, the feature behaved as if both were
hard-coded to 1. Some hard-coded English texts were turned into proper
strings, too.